+++ manifest.json
@@ -1,9 +1,9 @@
 {
 	"id": "tag-many",
 	"name": "TagMany",
-	"version": "0.2.4",
+	"version": "0.2.5",
 	"minAppVersion": "0.15.0",
-	"description": "This plugin allows you to add the same tag(s) to multiple notes in a folder (optionally including subfolders) at once.",
+	"description": "Add the same tag(s) to multiple notes in a folder (optionally including subfolders) at once.",
 	"author": "Joshua Martius",
 	"authorUrl": "https://fngt.nl/joshua",
 	"isDesktopOnly": true
